post policy

+++ app/Http/Controllers/Member/BlogController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BlogController extends Controller
 {
@@ -80,6 +81,7 @@ class BlogController extends Controller
      */
     public function edit(Post $post)
     {
+        Gate::authorize('edit',$post);
         $data = $post;
         return view('member.blogs.edit', compact('data'));
     }
@@ -129,7 +131,8 @@ class BlogController extends Controller
      */
     public function destroy(Post $post)
     {
-        Post::where('id', $post->id);
+        Gate::authorize('delete',$post);
+        
         if (isset($post->thumbnail) && file_exists(public_path(getenv('CUSTOM_THUMBNAIL_LOCATION')) . "/" . $post->thumbnail)) {
             unlink(public_path(getenv('CUSTOM_THUMBNAIL_LOCATION')) . "/" . $post->thumbnail);
         }
